Force encoding on invalid encoding test case
Ruby 1.9 tests were failing because the default encoding is US-ASCII
and the string was valid under that encoding. Ruby 2+ tests were ok
because the default encoding is now UTF-8. The solution here is to
force the encoding on the string in order to ensure that the string in
not a valid encoding for all cases.