Remove minitest-rails
There is currently no version of minitest-rails that is compatible with
Rails 7.2 or 8.0. Open pull requests in the repo are not being reviewed.

It seems we don't use the gem for anything other than adding minitest's
Spec::DLS to ActiveSupport::TestCase, and that we can just as easily do
ourselves.